Clean up implementation of Signature Status and DllCharacteristics
