+++ Lib/Objects/FileSystemObject.cs
@@ -17,26 +17,13 @@ namespace AttackSurfaceAnalyzer.ObjectTypes
 {
     public class FileSystemObject
     {
-        public static List<string> SIGNED_EXTENSIONS = new List<string> { "dll", "exe", "cab", "ocx" };
 
         public string Path;
         public string Permissions;
         public ulong Size;
         public string ContentHash;
-
-        public string SignatureStatus
-        {
-            get
-            {
-                if (!NeedsSignature())
-                {
-                    return "No signature required";
-                }
-                string sigStatus = WinTrust.VerifyEmbeddedSignature(Path);
-
-                return sigStatus;
-            }
-        }
+        public List<DLLCHARACTERISTICS> Characteristics;
+        public string SignatureStatus;
 
         public string RowKey
         {
@@ -46,77 +33,11 @@ namespace AttackSurfaceAnalyzer.ObjectTypes
             }
         }
 
-        public bool NeedsSignature()
-        {
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                var _p = Path.ToLower().Trim();
-                foreach (var ext in SIGNED_EXTENSIONS)
-                {
-                    if (_p.EndsWith("." + ext))
-                    {
-                        return true;
-                    }
-                }
-                return false;
-            }
-            else
-            {
-                return false;
-            }
-        }
-
-        public List<string> DllCharacteristics
-        {
-            get
-            {
-                if (!NeedsSignature())
-                {
-                    return new List<string>();
-                }
-                try
-                {
-                    List<string> l = new List<string>();
-                    var peHeader1 = new PeNet.PeFile(this.Path);
-                    ushort characteristics = peHeader1.ImageNtHeaders.OptionalHeader.DllCharacteristics;
-                    foreach (DLLCHARACTERISTICS characteristic in Enum.GetValues(typeof(DLLCHARACTERISTICS)))
-                    {
-                        if (((ushort)characteristic & characteristics) == (ushort)characteristic)
-                        {
-                            l.Add(characteristic.ToString());
-                        }
-                    }
-                    return l;
-                }
-                catch (Exception e)
-                {
-                    Log.Debug("{0}:{1}", e.GetType().ToString(), e.Message);
-                    Log.Debug(e.StackTrace);
-                }
-                return new List<string>();
-            }
-        }
+        
 
         public override string ToString()
         {
             return string.Format("Path={0}, Permission={1}, Size={2}, ContentHash={3}", Path, Permissions, Size, ContentHash);
-        }
-
-        public string ToJson()
-        {
-            var _path = JsonConvert.ToString(Path);
-            var _permission = JsonConvert.ToString(Permissions);
-            var _size = JsonConvert.ToString(Size);
-            var _contentHash = JsonConvert.ToString(ContentHash);
-
-            var sb = new StringBuilder();
-            sb.Append("{");
-            sb.AppendFormat("\"path\":{0},", _path);
-            sb.AppendFormat("\"permission\":{0},", _permission);
-            sb.AppendFormat("\"size\":{0},", _size);
-            sb.AppendFormat("\"content_hash\":{0}", _contentHash);
-            sb.AppendLine("}");
-            return sb.ToString();
         }
     }
 }
